+++ src/main/clip/LinesClip.java
