+++ src/main/environment/StrandPanel.java
